Moves code that imports Sinan to it's own file
